// src/interfaces/farm-summary.interface.ts
export interface FarmSummary {
    farmName: string;
    acAddress: string;
    stakingAddress: string;
    rewardsUSDPerYear: string;
    tvlUSD: number;
    userTlvUSD: number;
    token0Address: string;
    token1Address: string;
    stakingToken: string;
    tvl: number;
    apr: number;
    apy: number;
    tokenLogos: string[];
    userRewards: number;
}